🏗️ refactor [#10.3.8]: 테스트 3차 개선 - Fixture 중앙화, ALL_COMPLETED 패턴
🔍 변경 사항:
- **Fixture 중앙화**: `tests/conftest.py`의 [conftest.py]에 Obsidian sync 관련 fixture 추가
  - 모든 테스트 파일에서 중복 제거
  - [mock_vault], [obsidian_config], [sync_service], [map_manager]
- **동시성 테스트 개선**:
    - `return_when=ALL_COMPLETED` 사용 (FIRST_EXCEPTION 제거)
    - `len(done) == num_workers` 검증으로 명확한 완료 확인
    - Timeout과 예외 케이스 구분 명확화
- **코드 품질**: Import 정리, Docstring 개선, 주석 추가

📁 파일:
- tests/conftest.py (Modified)
- tests/unit/test_sync_map_manager.py (Modified)
- tests/unit/test_obsidian_sync_unit.py (Modified)
- tests/integration/test_obsidian_sync.py (Modified)

📌 Related
- Issue [#76]
- @ai-bot-review (https://github.com/jjaayy2222/flownote-mvp/pull/133#issuecomment-3625327007)
